Refactor services, a-propos, contact, and devis pages with unified style
