+++ pages_refaites.txt
@@ -1,0 +1,7 @@
+=== PAGES REFAITES AVEC LE STYLE DE L'ACCUEIL ===
+✅ services.php - Refait avec le style de l'accueil
+✅ a-propos.php - Refait avec le style de l'accueil
+✅ contact.php - Refait avec le style de l'accueil
+✅ devis.php - Refait avec le style de l'accueil
+
+📊 PROGRÈS : 4/20 pages refaites (20%)
